Expense average of room and expenses status of user in a room

+++ app/src/main/java/com/ltei/kunzmznzger/models/Expense.kt
@@ -10,7 +10,7 @@ import java.io.Serializable
 class Expense : Model<Expense>(), Serializable {
 
     var name: String? = null
-    var value: Float? = null
+    var value: Double? = null
     var description: String? = null
 
     var createdAt: DateTime? = null
